% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download.R
\name{dst_download}
\alias{dst_download}
\title{download data sets from Statistics Denmark}
\usage{
dst_download(tableID, vars, lang = "en")
}
\arguments{
\item{tableID}{the data set's ID. See dst_tables() to get the ID.}

\item{vars}{the variables to return. If not specified it will return time and value of the
what the data set measure.}

\item{lang}{used language in the data set. Can take the values "da" for danish and "en" for
english}
}
\description{
You can use this function to retrieve data sets from Statistics Denmark. It also let you choose
which variables to return.
}
\examples{
dst_download("folk1a")
dst_download("folk1a", c("ALDER", "CIVILSTAND"))
}
