% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{dst_information}
\alias{dst_information}
\title{get information about Statistics Denmark's data sets}
\usage{
dst_information(tableID, lang = "en", columns = c("id", "description",
  "contactperson", "mail", "url"))
}
\arguments{
\item{tableID}{the data set's ID. See dst_tables()}

\item{lang}{used language in the data set. Can take the values "da" for danish and "en" for
english.}

\item{columns}{the columns returned. columns can take the values: id, text, description,
unit, updated, name, phone, mail, documentationID, url or footnote.}
}
\description{
You can use the function to retrieve information about all the data sets available at
Statistics Denmark. For instance, you could get the time at which a specific data set
previously was updated
}
\examples{
dst_information("folk1a")
}
