% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjects.R
\name{dst_subjects}
\alias{dst_subjects}
\title{Retrieve subjects from Statistics Denmark}
\usage{
dst_subjects(lang = "en", columns = c("id", "description"))
}
\arguments{
\item{lang}{description language - "da" for danish and "en" for english.}

\item{columns}{columns to return. Can take the values "id", "description", "active",
"hasSubject" or "subjects".}
}
\value{
Tibble containing the specified columns in the chosen language.
}
\description{
You can use `dst_subject()` to get the subjects, which Statistics Denmark use to categorize
their data sets.
}
\examples{
dst_subjects()
dst_subjects(columns = c("id", "description", "active")) # is the data set active
}
