% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{dst_tables}
\alias{dst_tables}
\title{retrieve tables from Statistics Denmark}
\usage{
dst_tables(subjectsID, lang = "en", columns = c("id", "text"))
}
\arguments{
\item{subjectsID}{Should the data sets be within a specific field. See dst_subject() for IDs.}

\item{lang}{the desired language. Can take the values "da" for danish and "en" english.}

\item{columns}{the returned columns. Can take the values "id", "text", "unit", "updated",
"firstPeriod", "latestPeriod", "active" or "variables"}
}
\value{
Tibble which contain information about all tables available at Danish Statistics. If
subject is provided only tables within the subject is returned. tibble will contain the columns
specified in the columns argument. The language is as default English otherwise "da" is
provided in the lang argument.
}
\description{
You can use `dst_tables()` to get a tibble, which contain information about the data sets
available at the Bureau.
}
\examples{
dst_tables() # retrieve all tables
dst_tables(subjectsID = "02") # Population and elections data sets
}
