% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{dst_variable_values}
\alias{dst_variable_values}
\title{get tibble with variable values}
\usage{
dst_variable_values(tableID, lang = "en")
}
\arguments{
\item{tableID}{the data set's ID. See dst_tables()}

\item{lang}{used language in the data set. Can take the values "da" for danish and "en" for
english.}
}
\description{
you can use this function to see, which values the data set's variables can take
}
\examples{
dst_variable_values("folk1a")
}
