% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download.R
\name{dst_variables}
\alias{dst_variables}
\title{get data set variables}
\usage{
dst_variables(tableID, lang = "en", columns = c("id", "text"))
}
\arguments{
\item{tableID}{the id of the desired table. See dst_tables() to get an overview of the table IDs.}

\item{lang}{the language used in the description of the variables. "en" for english "da" for
danish.}

\item{columns}{the columns returned. columns can take the following values: id, text,
elimination, time or map}
}
\description{
You can use this function to get a tibble which has information about the available variables
in the respective data set as well as an description of those
}
\examples{
## Not run:
dst_download("FOLK1A")
## End(Not run)
dst_variables("FOLK1A", columns = c("id", "text", "time"))
}
